solve bug and improve parallelization
There was an unstable behaviour due to the fact that clusterExport() is taking variables in global environment by default. I have clarified it in function parCompute.
In order to improve the parallelization, I have shuffled the list of K (number of bootstrap samples) so that there is roughly the same computation time in each thread.